+++ plugins/FavContacts/src/menu.cpp
@@ -47,7 +47,7 @@ static BOOL sttMeasureItem_Group(LPMEASUREITEMSTRUCT lpmis, Options *options)
 		SelectObject(hdc, g_Options.hfntName);
 
 	SIZE sz;
-	GetTextExtentPoint32(hdc, name, mir_tstrlen(name), &sz);
+	GetTextExtentPoint32(hdc, name, (int)mir_tstrlen(name), &sz);
 	lpmis->itemHeight = sz.cy + 8;
 	lpmis->itemWidth = sz.cx + 10;
 	SelectObject(hdc, hfntSave);
@@ -81,7 +81,7 @@ static BOOL sttMeasureItem_Contact(LPMEASUREITEMSTRUCT lpmis, Options *options)
 
 		if (!options->bSysColors)
 			SelectObject(hdc, g_Options.hfntSecond);
-		GetTextExtentPoint32(hdc, title, mir_tstrlen(title), &sz);
+		GetTextExtentPoint32(hdc, title, (int)mir_tstrlen(title), &sz);
 		textWidth = sz.cx;
 		lpmis->itemHeight += sz.cy + 3;
 
@@ -92,7 +92,7 @@ static BOOL sttMeasureItem_Contact(LPMEASUREITEMSTRUCT lpmis, Options *options)
 	TCHAR *name = (TCHAR *)CallService(MS_CLIST_GETCONTACTDISPLAYNAME, hContact, GCDNF_TCHAR);
 
 	if (!options->bSysColors) SelectObject(hdc, g_Options.hfntName);
-	GetTextExtentPoint32(hdc, name, mir_tstrlen(name), &sz);
+	GetTextExtentPoint32(hdc, name, (int)mir_tstrlen(name), &sz);
 	textWidth = max(textWidth, sz.cx);
 
 	SelectObject(hdc, hfntSave);
@@ -309,7 +309,7 @@ static BOOL sttDrawItem_Contact(LPDRAWITEMSTRUCT lpdis, Options *options = NULL)
 		DrawText(hdcTemp, name, -1, &lpdis->rcItem, DT_NOPREFIX | DT_SINGLELINE | DT_TOP | DT_LEFT);
 
 		SIZE sz;
-		GetTextExtentPoint32(hdcTemp, name, mir_tstrlen(name), &sz);
+		GetTextExtentPoint32(hdcTemp, name, (int)mir_tstrlen(name), &sz);
 		lpdis->rcItem.top += sz.cy + 3;
 	}
 
@@ -390,8 +390,8 @@ static LRESULT CALLBACK MenuHostWndProc(HWND hwnd, UINT message, WPARAM wParam, 
 			RemoveMenu((HMENU)lParam, 1, MF_BYPOSITION);
 
 		if (LOWORD(wParam) == VK_BACK) {
-			if (int l = mir_tstrlen(g_filter))
-				g_filter[l - 1] = 0;
+			if (size_t l = mir_tstrlen(g_filter))
+				g_filter[l-1] = 0;
 		}
 		else if (_istalnum(LOWORD(wParam))) {
 			if (mir_tstrlen(g_filter) < SIZEOF(g_filter) - 1) {
